refactor: Rename the `build_quiz` into `prepare` to avoid quiz repetition in the folders
