+++ lib/src/quiz/prepare/prepare_quiz_view.dart
@@ -2,23 +2,23 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:go_router/go_router.dart';
 import 'package:kana_to_kanji/src/core/widgets/app_scaffold.dart';
-import 'package:kana_to_kanji/src/quiz/build_quiz/build_quiz_view_model.dart';
+import 'package:kana_to_kanji/src/quiz/prepare/prepare_quiz_view_model.dart';
 import 'package:kana_to_kanji/src/core/constants/alphabets.dart';
-import 'package:kana_to_kanji/src/quiz/build_quiz/widgets/alphabet_groups_expansion_tile.dart';
+import 'package:kana_to_kanji/src/quiz/prepare/widgets/alphabet_groups_expansion_tile.dart';
 import 'package:kana_to_kanji/src/core/widgets/chip_list.dart';
 import 'package:stacked/stacked.dart';
 
-class BuildQuizView extends StatelessWidget {
+class PrepareQuizView extends StatelessWidget {
   static const routeName = "/quiz";
 
-  const BuildQuizView({super.key});
+  const PrepareQuizView({super.key});
 
   @override
   Widget build(BuildContext context) {
     final l10n = AppLocalizations.of(context);
 
-    return ViewModelBuilder<BuildQuizViewModel>.reactive(
-      viewModelBuilder: () => BuildQuizViewModel(GoRouter.of(context)),
+    return ViewModelBuilder<PrepareQuizViewModel>.reactive(
+      viewModelBuilder: () => PrepareQuizViewModel(GoRouter.of(context)),
       builder: (context, viewModel, _) => AppScaffold(
           appBar: AppBar(
             title: Text(l10n.quiz_build_title),
