+++ lib/src/quiz/prepare/widgets/alphabet_groups_expansion_tile.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:kana_to_kanji/src/core/constants/alphabets.dart';
 import 'package:kana_to_kanji/src/core/models/group.dart';
-import 'package:kana_to_kanji/src/quiz/build_quiz/widgets/group_card.dart';
-import 'package:kana_to_kanji/src/quiz/build_quiz/widgets/kana_groups.dart';
+import 'package:kana_to_kanji/src/quiz/prepare/widgets/group_card.dart';
+import 'package:kana_to_kanji/src/quiz/prepare/widgets/kana_groups.dart';
 
 class AlphabetGroupsExpansionTile extends StatelessWidget {
   final Alphabets alphabet;
