+++ lib/src/quiz/prepare/widgets/kana_groups_section.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:kana_to_kanji/src/core/constants/kana_type.dart';
 import 'package:kana_to_kanji/src/core/models/group.dart';
-import 'package:kana_to_kanji/src/quiz/build_quiz/widgets/group_card.dart';
+import 'package:kana_to_kanji/src/quiz/prepare/widgets/group_card.dart';
 
 class KanaGroupsSection extends StatelessWidget {
   final KanaTypes type;
